Finish basic parser
I've built a pretty simple vector structure to verify my parser works.
It simply holds a sequence of tokens arranged in a way that roughly
corresponds to that of reverse polish notation. This mirrors the
structure that the bytecode is going to have, since the virtual machine
is going to be stack-based.

After some testing using this data structure, I'm pretty sure the parser
is fully functional. It handles arithmetic expressions like a charm. Now
I'll start working on the bytecode, using the vector thing I built as a
base.