Add bytecode data structure
It is basically a dynamic array of bytes and a dynamic array of cog
values. The array of bytes contains the code, and the array of values is
a constant pool referenced by the code.

I've named the structure a "box" for the lack of a better, shorter
name. It also kind of plays to the aesthetic estabilished by "cog",
as it is another factory thing, if that makes sense.