Add basic virtual machine
Finally, an overengineered calculator! The error reporting is still very
lacking, and it supports a very minimal subset of the language I want to
build, but it works. You give it an expression, it gives a result back.
Internally, it does lexing, parsing, bytecode compiling and bytecode
processing.

In conclusion, I think I can now say I have a pretty decent base for a
complex interpreter.